modify portfoliopage component design and add fetching of current prices
