+++ stock-tracker/package.json
@@ -32,6 +32,8 @@
     "postcss": "8.4.24",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-loader-spinner": "^5.4.5",
+    "react-spinners": "^0.13.8",
     "recharts": "^2.8.0",
     "styled-components": "^5.3.11",
     "tailwind-merge": "^1.14.0",
