+++ stock-tracker/src/app/components/portfolio/PortfolioGrid.module.css
@@ -1,0 +1,45 @@
+.grid{
+    display: grid;
+    justify-content: center;
+    margin: 1rem;
+    padding: 1rem;
+
+    border: 1px solid ;
+    
+    border-radius: 12px;
+    
+    width: 80%;
+}
+
+@media (prefers-color-scheme: dark){
+    .grid{
+        border-color: aliceblue;
+        color: aliceblue;
+    }
+    .MuiPaginationItem-root {
+        color: white !important; /* Change to dark mode's desired color */
+    }
+    .grid_header{
+        color:aliceblue;
+        font-weight: bold;
+        font-size: 1.2rem;
+    }
+}
+.grid_header{
+    color:black;
+    font-weight: bold;
+    font-size: 1.2rem;
+}
+.grid.MuiDataGrid-columnHeaderTitle {
+    font-weight: bold;
+    font-size: 2.5rem;
+}
+
+.negativeReturn{
+    color: red;
+    font-weight: bold;
+}
+.positiveReturn{
+    color: green;
+    font-weight: bold;
+}
