+++ stock-tracker/src/app/portfolio/LoadingPage.module.css
@@ -1,0 +1,17 @@
+.loader_container {
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    position: fixed;
+}
+
+.loader_container {
+    width: 100%;
+    height: 100vh;
+    position: fixed;
+    background: url("https://media.giphy.com/media/8agqybiK5LW8qrG3vJ/giphy.gif") center no-repeat;
+    z-index: 1;
+}
+
