+++ stock-tracker/src/app/portfolio/PortfolioPage.module.css
@@ -1,46 +1,4 @@
-.grid{
-    display: grid;
-    justify-content: center;
-    margin: 1rem;
-    padding: 1rem;
 
-    border: 1px solid ;
-    
-    border-radius: 12px;
-    
-    width: 80%;
-}
-
-@media (prefers-color-scheme: dark){
-    .grid{
-        border-color: aliceblue;
-        color: aliceblue;
-    }
-    .MuiPaginationItem-root {
-        color: white !important; /* Change to dark mode's desired color */
-    }
-    .grid_header{
-        color: aliceblue;
-    }
-}
-.grid_header{
-    color:black;
-    font-weight: bold;
-    font-size: 1.2rem;
-}
-.grid.MuiDataGrid-columnHeaderTitle {
-    font-weight: bold;
-    font-size: 2.5rem;
-}
-
-.negativeReturn{
-    color: red;
-    font-weight: bold;
-}
-.positiveReturn{
-    color: green;
-    font-weight: bold;
-}
 .header{
     padding: 1%;
     font-weight: bold;
@@ -51,40 +9,4 @@
     justify-content: center;
     align-items: center;
     flex-direction: column;
-}
-.loader_container {
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: fixed;
-    z-index: 1;
-}
-
-.loader_container {
-    width: 100%;
-    height: 100vh;
-    position: fixed;
-    background: url("https://media.giphy.com/media/8agqybiK5LW8qrG3vJ/giphy.gif") center no-repeat;
-    z-index: 1;
-}
-
-.spinner {
-    width: 64px;
-    height: 64px;
-    border: 8px solid;
-    border-color: #00c3ff transparent #00c3ff transparent;
-    border-radius: 50%;
-    animation: spin-anim 1.2s linear infinite;
-}
-
-@keyframes spin-anim {
-    0% {
-        transform: rotate(0deg);
-    }
-
-    100% {
-        transform: rotate(360deg);
-    }
 }
